requirements.txt
pytest==7.4.0
constructs>=10.0.0,<11.0.0
aws-cdk-lib==2.54.0


